rename particle_seeding to particle_seeder
