+++ src/sedtrails/particle_tracer/particle_seeder.py
